requirements.txt
allure-pytest==2.15.3
email_validator==2.3.0
playwright==1.55.0
pydantic==2.12.5
pydantic-settings==2.13.0
pytest==9.0.1
pytest-playwright==0.7.2
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
ui-coverage-tool==0.33.0